update remove script

+++ remove-all.sh
@@ -4,12 +4,13 @@ project_path=$(cd `dirname $0`; pwd)/
 
 cd ${project_path}
 
-
-cd  SenseCAP-DataVirt-SaaS-Example/
-docker-compose stop
-docker-compose kill
-docker-compose rm
+if [ -d "/SenseCAP-DataVirt-SaaS-Example" ]; then
+    cd  SenseCAP-DataVirt-SaaS-Example
+    docker-compose stop
+    docker-compose kill
+    docker-compose rm
+fi
 
 
 cd ${project_path}
-rm -rf SenseCAP-DataVirt-SaaS-Example/
+rm -rf SenseCAP-DataVirt-SaaS-Example
